Shifting to shared build_totals
